backend-py/requirements.txt
fastapi
uvicorn[standard]
pywifi
comtypes
speedtest-cli
google-generativeai
python-dotenv
apscheduler
pydantic
requests
manuf